Создал модульность в проекте, переписал функцию - исправив ошибку повторяющегося id и url, подключил модуль в index.html
